requirements/dev.txt
django-extensions==3.1.5
Faker==9.8.2